+++ typescript/src/cairoToMidiParser.ts
@@ -2,35 +2,27 @@ export interface CairoParsedMidiEvent {
     type: string;
     channel: number;
     noteNumber: number;
-    velocity: number;
-    deltaTime: number;
-    microsecondsPerBeat: number;
-    numerator: number;
+    velocity: number; //how 'hard' the note is pressure. If == 0 it's and noteOff event. But a noteOff can be != 0 too.
+    deltaTime: number; //distance in between the event n and the event n - 1
+    microsecondsPerBeat: number; //the tempo
+    numerator: number; // with [numerador/denominator] you can generate the compass
     denominator: number;
     value: number;
     pitch: number;
     clock: number;
-    absoluteTime: number;
+    absoluteTime: number; // events in cairo don't have the absolute time (they only have the deltaTime), and in @tonejs/midi yes. It's required to calculate it
     controllerType: number;
-    ticksPerBeat: number;
-    meta: boolean;
-    programNumber: number;
+    ticksPerBeat: number; ////the same as ppq
+    meta: boolean; //if the event it's for metadata
+    programNumber: number; //required to change instrument type, for example 24 = Nylon String Guitar
 }
-
-function toCamelCase(str: string) {
-    return str
-        .toLowerCase() // Convertir a minúsculas
-        .split('_')    // Dividir por el guion bajo
-        .map((word, index) =>
-            index === 0
-                ? word
-                : word.charAt(0).toUpperCase() + word.slice(1) // Capitalizar la primera letra desde el segundo elemento
-        )
-        .join(''); // Unir las palabras sin espacios
-}
-
-export function parseEvent(line: string): CairoParsedMidiEvent | null {
-    const match = line.match(/Message::([A-Z_]+)\((.+)\)/);
+/**
+ * Parse a cairo event into a CairoParsedMidiEvent.
+ * CairoParsedMidiEvent contains all possibles properties for a MidiEvent of the library @tonejs/midi
+ * the name of the properties are the same of that library for compatibility
+ */
+export function parseEvent(cairoEvent: string): CairoParsedMidiEvent | null {
+    const match = cairoEvent.match(/Message::([A-Z_]+)\((.+)\)/);
     if (!match) return null;
 
     const [, type, content] = match;
@@ -47,6 +39,7 @@ export function parseEvent(line: string): CairoParsedMidiEvent | null {
             parsedEvent.channel = parseInt(content.match(/channel: (\d+)/)?.[1] || "0");
             parsedEvent.noteNumber = parseInt(content.match(/note: (\d+)/)?.[1] || "0");
             parsedEvent.velocity = parseInt(content.match(/velocity: (\d+)/)?.[1] || "0");
+            // This is based on midi specification
             if (parsedEvent.velocity === 0) {
                 parsedEvent.type = 'noteOff'
             }
@@ -61,7 +54,7 @@ export function parseEvent(line: string): CairoParsedMidiEvent | null {
             if (timeOptionMatch) {
                 parsedEvent.deltaTime = parseFP32x32(timeOptionMatch[1]);
             } else {
-                parsedEvent.deltaTime = undefined;  // No `time` provided
+                parsedEvent.deltaTime = undefined;
             }
             parsedEvent.meta = true
             break;
@@ -109,10 +102,22 @@ export function parseEvent(line: string): CairoParsedMidiEvent | null {
     return parsedEvent as CairoParsedMidiEvent;
 }
 
+function toCamelCase(str: string) {
+    return str
+        .toLowerCase()
+        .split('_')
+        .map((word, index) =>
+            index === 0
+                ? word
+                : word.charAt(0).toUpperCase() + word.slice(1)
+        )
+        .join('');
+}
+
 function parseFP32x32(fp: string): number  {
     const match = fp.match(/FP32x32 { mag: (\d+), sign: (\w+) }/);
     if (match) {
-        return parseInt(match[1], 10);  // Solo el valor 'mag' como número
+        return parseInt(match[1], 10);
     }
     return 0;
 }
